// VIS_Application/app/Model/Masters/CompanyRelated/ContactMaster.ts
export interface IContactMaster {
    
    Id: number;
    Name: string;
    Designation: string;
    Email: string;
    phone: string;
    SkypeId: string;
    MsnId: string;
    GtalkId: string;
    AolId: string;
    Other: string;
    ProspectId: number;
    CreatedOn: Date;
    CreatedBy: string;
    UpdatedOn: Date;
    UpdatedBy: string
    IsActive: boolean;

}